requirements.txt
absl-py==1.4.0
ansi2html==1.8.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.3.0
chart-studio==1.1.0
click==8.1.3
cloudpickle==2.2.1
colorlover==0.3.0
cufflinks==0.17.3
Cython==0.29.34
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
deprecation==2.1.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
jax==0.4.8
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.39.1
Markdown==3.4.3
matplotlib==3.7.1
missingno==0.4.2
ml-dtypes==0.0.4
munkres==1.1.4
numba==0.56.4
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.9
pandas==1.5.3
Pillow==9.4.0
ply==3.11
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyod==1.0.9
pyparsing==3.0.9
PyQt5-sip==12.11.0
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.9
schemdraw==0.16
scikit-plot==0.3.7
scipy==1.10.1
sktime==0.17.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.2.0
tqdm==4.65.0
trace-updater==0.0.9
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
wurlitzer==3.0.3
xgboost==1.7.5
xxhash==3.2.0
yellowbrick==1.5
zipp==3.15.0
